requirements.txt
alembic==1.4.3
black==20.8b1
Flask==1.1.2
google-cloud-storage==1.31.2
isort==5.5.3
psycopg2==2.8.6
SQLAlchemy==1.3.19
